+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+## Added
+
+- Add a retry option for the initial connection probing. This is disabled per
+  default but enabled in the Dockerfile.
+
+
 ## [0.8.2] - 2024-12-01
 
 ## Fixed